move client side renderer to pro directory
